requirements.txt
python-dotenv==0.19.2
flask==2.0.2
flask_recaptcha
flask_restful==0.3.9
numpy==1.21.2
pandas==1.4.1
pillow==9.0.1
